Implementação de id incremental na requisição de mudança

+++ src/main/java/br/gov/caixa/sammp/repository/RequisicaoMudancaRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 
 import br.gov.caixa.sammp.entity.RequisicaoMudanca;
 
-public interface RequisicaoMudancaRepository extends CrudRepository<RequisicaoMudanca, Double>{
+public interface RequisicaoMudancaRepository extends CrudRepository<RequisicaoMudanca, Long>{
 
 	@Query("select rm from RequisicaoMudanca rm where "
 			+ "(rm.dataEntrega between :dataInicio and :dataFim) or "
